Fix the display of elements on the authentication and registration pages and decomposed some of the code
