+++ src/pages/auth/authLayout/AuthLayout.module.css
@@ -1,14 +1,40 @@
-.loginStyleH1 {
+.styleH1 {
     font-size: 32px;
     text-align: center;
 }
 
-.loginStyleH2 {
+.styleH2 {
     font-size: 26px;
     text-align: center;
 }
 
-.regStyleH1 {
-    font-size: 32px;
-    text-align: center;
+.mainTitleContainer {
+    @apply block justify-center items-center mb-4;
+}
+
+.titleLayout {
+    @apply flex justify-center items-center;
+}
+
+.titleSize {
+    font-size: 26px;
+}
+
+.inputsBox {
+    @apply pl-5;
+}
+
+.authContainer {
+    @apply min-h-max min-w-[300px] flex w-[min(80%,60rem)] flex-col flex-nowrap items-center justify-center gap-6 self-center bg-[radial-gradient(165.81%_165.81%_at_-13.5%_-44.62%,rgba(255,255,255,0.1)_0%,rgba(255,255,255,0)_100%)] rounded-[20px] pb-14 pt-14 mb-20;
+}
+
+.referenceBack {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+.textError {
+    /*If we decide to return the inscription*/
+    @apply text-[18px] text-[#B52D56]
 }
