--- src/pages/auth/login/Login.module.css
@@ -1,4 +1,0 @@
-.titleLayout {
-    font-size: 26px;
-}
-
