--- src/pages/auth/registration/Registration.module.css
@@ -1,3 +1,0 @@
-.titleLayout {
-    font-size: 26px;
-}
